requirements.txt
pandas
numpy
requests
plotly
kaleido
python-dotenv
google-genai